+++ lib/back_stack.dart
@@ -7,6 +7,7 @@
 library;
 
 export 'src/nav_app.dart';
+export 'src/nav_async_redirect.dart';
 export 'src/nav_confirm.dart';
 export 'src/nav_display.dart';
 export 'src/nav_entries.dart';